refactor: upgrade to react-leaflet v2
