+++ packages/ui/package.json
@@ -27,7 +27,7 @@
     "react": "16.4.1",
     "react-addons-update": "15.6.2",
     "react-dom": "16.4.1",
-    "react-leaflet": "1.9.1",
+    "react-leaflet": "2.0.0",
     "react-redux": "5.0.7",
     "react-responsive": "4.1.0",
     "react-router-dom": "4.3.1",
